optimize: reduce the granularity of lock in method refreshLine() and initAndSendLinesInfo().
in the past code, the two methods were synchronized by using ReentrantLock to make the line data consistent, and it works well, but the lock granularity is coarse. it shows low efficiency when handling multiple requests of refreshLine() and there is unnecessary to lock in this case.

the demand is that locking when calling initAndSendLinesInfo() and not at any other time.

reconstruct it by the way