update: rate limit logic
The previous way of adding iptables rules one by one will lead to confusion of rules for unknown reason.

The possible reason is represented at the comment of class Process says 'The methods that create processes may not work well for special processes on certain native platforms, such as native windowing processes, daemon processes, Win16/DOS processes on Microsoft Windows, or shell scripts.'

so I replace the old rule creation way with 'iptables-restore',which uses a mechanism similar to TRANSACTION to prevent confusion.

But sometimes it still doesn't work. It shows that nothing has been created in the rule table.