fix: createLine() called in refresh is no aspect weaved
when refresh() is called, the creatingLine set is not updated, and the lineMonitor will create the line multi times with no ip address returned in the dialing process at fast speed. so the  line will be sent as deadLine before refresh() is done, if the dialing process cost too much times.

Taking one time dialing false line for deadline is not obey the rule of deadLine.